Polish call replacers.. These aren't used anywhere thus far. Also refactor several helper methods into utils module.
